rework match.run and match end for n players
