+++ scenes/Game2pVs.lua
@@ -19,7 +19,26 @@ Game2pVs.name = "Game2pVs"
 sceneManager:addScene(Game2pVs)
 
 function Game2pVs:processGameResults(gameResult)
-  Replay.finalizeAndWriteVsReplay(nil, false, self.match, replay)
+  local extraPath, extraFilename = "", ""
+  local rep_a_name, rep_b_name = self.match.players[1].name, self.match.players[2].name
+    --sort player names alphabetically for folder name so we don't have a folder "a-vs-b" and also "b-vs-a"
+    if rep_b_name < rep_a_name then
+      extraPath = rep_b_name .. "-vs-" .. rep_a_name
+    else
+      extraPath = rep_a_name .. "-vs-" .. rep_b_name
+    end
+    extraFilename = extraFilename .. rep_a_name .. "-L" .. self.match.P1.level .. "-vs-" .. rep_b_name .. "-L" .. self.match.P2.level
+    if match_type and match_type ~= "" then
+      extraFilename = extraFilename .. "-" .. match_type
+    end
+    if not self.match.replay.incomplete then
+      if self.match.replay.winnerIndex then
+        extraFilename = extraFilename .. "-P" .. self.match.replay.winnerIndex .. "wins"
+      else
+        extraFilename = extraFilename .. "-draw"
+      end
+    end
+  Replay.finalizeAndWriteReplay(extraPath, extraFilename, self.match.replay)
 end
 
 return Game2pVs
